rust: recompiled protobuf, renamed modules
