+++ rust/src/lib.rs
@@ -1,10 +1,10 @@
+mod api;
 mod helpers;
 mod ics23;
 mod ops;
-mod proofs;
 mod verify;
 
-pub use crate::proofs::*;
+pub use api::{iavl_spec, tendermint_spec, verify_membership, verify_non_membership};
 pub use helpers::{Hash, Result};
-pub use crate::ics23::{iavl_spec, tendermint_spec, verify_membership, verify_non_membership};
+pub use crate::ics23::*;
 pub use verify::calculate_existence_root;
